Adds custom ui and a TON of code
The gun now cycles the belts properly, ejecting empty ones and loading new ones from the hopper as expected. The gun stalls out when firing without a hopper which can waste ammo, functioning ui thanks to @Bobbanz1 works as expected for the most part but need some more pieces added to it when the gun is stalled or the cycler is jammed, cycler durability lowers with every shot. There's probably a lot more in here but I added way too much to keep track of in a commit description, reminder to make smaller commits